BackEnd/ExecuteServer/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-import-export==3.2.0
django-redis==5.4.0
django-storages==1.14.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==7.0.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
kombu
MarkupPy==1.14
# odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
ply==3.11
prompt-toolkit==3.0.39
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.7.0
xlrd==2.0.1
xlwt==1.3.0
python-dotenv
django-cors-headers==4.2.0